Fix for newer astropy/numpy
It was complaining that a unitless Quantity was being passed to sinc.
I'm not sure if that's due to upgrading astropy or because a newer numpy
implements NEP 18.

+++ katsdpimager/predict.py
@@ -58,7 +58,7 @@ def _extract_sky_model(image_parameters, grid_parameters, model, phase_centre):
     #
     # I'm not sure that this is mathematically defensible, but the results
     # look good.
-    taper = np.sinc(lmn[:, 0:2] / (ip.image_size * grid_parameters.oversample))
+    taper = np.sinc(lmn[:, 0:2] / float(ip.image_size * grid_parameters.oversample))
     flux *= np.product(taper, axis=1, keepdims=True)
 
     # For each image polarization, find the corresponding index in IQUV
